Extract SSO out of ApplicationController into its own module
